+++ test/testScripts/lightspeed/testLightspeed.test.ts
@@ -6,6 +6,8 @@ import {
   enableLightspeedSettings,
   disableLightspeedSettings,
   canRunLightspeedTests,
+  testInlineSuggestionNotTriggered,
+  testInlineSuggestionCursorPositions,
 } from "../../helper";
 
 function testSuggestionPrompts() {
@@ -20,6 +22,30 @@ function testSuggestionPrompts() {
     },
   ];
 
+  return tests;
+}
+
+function testInvalidPrompts() {
+  const tests = [
+    "-name: Print hello world",
+    "--name: Print hello world",
+    "- -name: Print hello world",
+    "-- name: Print hello world",
+  ];
+  return tests;
+}
+
+function testInvalidCursorPosition() {
+  const tests = [
+    {
+      taskName: "- name: Print hello world 1",
+      newLineSpaces: 2,
+    },
+    {
+      taskName: "- name: Print hello world 2",
+      newLineSpaces: 6,
+    },
+  ];
   return tests;
 }
 
@@ -53,8 +79,36 @@ export function testLightspeed(): void {
       });
     });
 
-    after(async function () {
-      disableLightspeedSettings();
+    describe("Test Ansible prompt not triggered", function () {
+      const docUri1 = getDocUri("lightspeed/playbook_1.yml");
+
+      before(async function () {
+        await vscode.commands.executeCommand(
+          "workbench.action.closeAllEditors"
+        );
+        await activate(docUri1);
+      });
+
+      const tests = testInvalidPrompts();
+
+      tests.forEach((promptName) => {
+        it(`Should not give inline suggestion for task prompt '${promptName}'`, async function () {
+          await testInlineSuggestionNotTriggered(promptName);
+        });
+      });
+
+      const invalidCursorPosTest = testInvalidCursorPosition();
+      invalidCursorPosTest.forEach(({ taskName, newLineSpaces }) => {
+        it(`Should not give inline suggestion for task prompt '${taskName}' with new line spaces ${newLineSpaces}`, async function () {
+          await testInlineSuggestionCursorPositions(
+            taskName,
+            newLineSpaces as number
+          );
+        });
+      });
+      after(async function () {
+        disableLightspeedSettings();
+      });
     });
   });
 }